feat: replace config-check endpoint with email-service-check
- Rename /api/config-check to /api/email-service-check for clearer purpose
- Use EmailServiceConfigurationValidator directly in endpoint
- Return simple {configured: boolean} response to frontend
- Log detailed validation problems server-side for debugging
- Endpoint now focuses specifically on email service configuration validation
